@@ -1,51 +1,58 @@
 {
   "name": "vue-click",
   "version": "1.5.1",
+  "private": false,
   "description": "Vue plugin for advanced click directive.",
+  "author": "Zachary Cardoza",
+  "scripts": {
+    "build": "webpack --config webpack/prod.config.js",
+    "lint": "vue-cli-service lint",
+    "dev": "webpack-dev-server --config webpack/dev.config.js",
+    "start": "npm run dev"
+  },
+  "main": "dist/index.js",
+  "files": [
+    "dist/index.js",
+    "dist/index.d.ts"
+  ],
+  "dependencies": {},
+  "devDependencies": {
+    "@vue/cli-plugin-eslint": "^4.5.11",
+    "@vue/eslint-config-standard": "^5.1.2",
+    "clean-webpack-plugin": "^3.0.0",
+    "dts-bundle": "^0.7.3",
+    "eslint": "^6.7.2",
+    "eslint-plugin-import": "^2.20.2",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "eslint-plugin-vue": "^7.0.0-0",
+    "html-webpack-plugin": "^3.2.0",
+    "ts-loader": "^6.2.1",
+    "tslint": "^6.0.0",
+    "tslint-config-standard": "^9.0.0",
+    "typescript": "^3.8.3",
+    "vue": "^3.0.7",
+    "webpack": "^4.42.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-dev-server": "^3.10.3"
+  },
+  "bugs": {
+    "url": "https://github.com/bayssmekanique/vue-click/issues"
+  },
+  "homepage": "https://github.com/bayssmekanique/vue-click#readme",
   "keywords": [
     "vue",
     "click",
     "throttle",
     "debounce"
   ],
-  "author": "Zachary Cardoza",
   "license": "MIT",
+  "publishConfig": {
+    "access": "public"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/bayssmekanique/vue-click.git"
   },
-  "bugs": {
-    "url": "https://github.com/bayssmekanique/vue-click/issues"
-  },
-  "homepage": "https://github.com/bayssmekanique/vue-click#readme",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "private": false,
-  "files": [
-    "dist/index.js",
-    "dist/index.d.ts"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "start": "npm run dev",
-    "lint": "tslint --project tsconfig.json",
-    "dev": "webpack-dev-server --config webpack/dev.config.js",
-    "build": "webpack --config webpack/prod.config.js"
-  },
-  "devDependencies": {
-    "clean-webpack-plugin": "^3.0.0",
-    "dts-bundle": "^0.7.3",
-    "html-webpack-plugin": "^3.2.0",
-    "ts-loader": "^6.2.1",
-    "tslint": "^6.0.0",
-    "tslint-config-standard": "^9.0.0",
-    "typescript": "^3.8.3",
-    "vue": "^2.6.11",
-    "webpack": "^4.42.0",
-    "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^3.10.3"
-  },
-  "dependencies": {}
+  "types": "dist/index.d.ts"
 }
